test(select): migrate card
