+++ core/src/components/select/test/card/select.e2e.ts-snapshots/select-card-md-ltr-Mobile-Chrome-linux.png
